pyproject.toml
[project]
name = "clean-backup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hachoir>=3.3.0",
    "pillow>=12.0.0",
    "pillow-heif>=1.1.1",
]
